# Electron/src/swagger/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spotify Electron Backend API",
    "description": "API created with Python FastAPI to serve          as backend for Spotify Electron music streaming Desktop App",
    "contact": {
      "name": "Antonio Mart\u00ednez Fern\u00e1ndez",
      "url": "https://github.com/AntonioMrtz",
      "email": "antoniomartinezfernandez17@gmail.com"
    },
    "license": {
      "name": "Attribution-NonCommercial-ShareAlike 4.0 International",
      "url": "https://creativecommons.org/licenses/by-nc-sa/4.0/deed.es"
    },
    "version": "2.1.0"
  },
  "paths": {
...
    "/users/whoami": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Who Am I",
        "description": "Returns token info from JWT\n\nArgs:\n    token: the jwt token. Defaults to None.",
        "operationId": "get_who_am_i_users_whoami_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "JWTBearer": []
          }
        ]
      }
    },
    "/users/{name}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User",
        "description": "Get user by name\n\nArgs:\n    name: user name\n    token: JWT info",
        "operationId": "get_user_users__name__get",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete User",
        "description": "Delete user\n\nArgs:\n    name: user name",
        "operationId": "delete_user_users__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create User",
        "description": "Create user\n\nArgs:\n    name: user name\n    photo: user photo\n    password: user password",
        "operationId": "create_user_users__post",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "photo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo"
            }
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{name}/playback_history": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Patch Playback History",
        "description": "Add song to playback history\n\nArgs:\n    name: user name\n    song_name: song name\n    token: JWT info",
        "operationId": "patch_playback_history_users__name__playback_history_patch",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "song_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Song Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Playback History",
        "description": "Get user song playback history\n\nArgs:\n    name: user name\n    token: JWT info",
        "operationId": "get_user_playback_history_users__name__playback_history_get",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{name}/saved_playlists": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Patch Saved Playlists",
        "description": "Add playlist to saved list\n\nArgs:\n    name: user name\n    playlist_name: saved playlist\n    token: JWT info",
        "operationId": "patch_saved_playlists_users__name__saved_playlists_patch",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "playlist_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete Saved Playlists",
        "description": "Delete playlist from saved list of user\n\nArgs:\n    name: user name\n    playlist_name: playlist name\n    token: JWT info",
        "operationId": "delete_saved_playlists_users__name__saved_playlists_delete",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "playlist_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{name}/relevant_playlists": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Relevant Playlists",
        "description": "Get relevant playlists for user\n\nArgs:\n    name: user name\n    token: JWT info",
        "operationId": "get_user_relevant_playlists_users__name__relevant_playlists_get",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{name}/playlists": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Playlists",
        "description": "Get playlists created by the user\n\nArgs:\n    name: user name\n    token: JWT info",
        "operationId": "get_user_playlists_users__name__playlists_get",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{name}/playlist_names": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Playlists Names",
        "description": "Get playlist names created by user\n\nArgs:\n    name: user name\n    token: JWT info",
        "operationId": "get_user_playlists_names_users__name__playlist_names_get",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{name}/promote": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Promote User To Artist",
        "description": "Promote user to artist\n\nArgs:\n    name: user name\n    token: JWT info",
        "operationId": "promote_user_to_artist_users__name__promote_patch",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "JWTBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}